+++ packages/dialtone-vue2/components/dropdown/dropdown.vue
@@ -478,7 +478,7 @@ export default {
     },
 
     onKeyPress (e) {
-      if (!this.isOpen || !this.isArrowKeyNav || !this.isValidCharacter(e.key)) {
+      if (!this.isOpen || !this.isArrowKeyNav || !this.isValidLetter(e.key)) {
         return;
       }
