+++ packages/dialtone-vue2/docs/component_driven_development/yeoman_generator.mdx
@@ -11,8 +11,8 @@ project.
 
 ```bash
 # From Dialtone Vue root directory
-npm run install:all
-npx yo dialtone-vue
+pnpm install
+pnpm exec yo dialtone-vue
 ```
 
 After running the command above you will be prompted for the name of the new component that you would like to create. It
