+++ packages/dialtone-vue2/docs/storybook/addons.stories.mdx
@@ -1,0 +1,10 @@
+import { Meta, Subtitle } from '@storybook/addon-docs';
+
+<Meta title="Docs/Storybook/Addons" />
+
+# Addons
+<Subtitle>
+  Storybook Addons and Configuration
+</Subtitle>
+
+`TODO`
