+++ packages/dialtone-vue2/docs/storybook/getting_started.mdx
@@ -22,7 +22,7 @@ Ensure that `NPM_AUTH_TOKEN` is set in your path. You will need to set the token
 ### Install Dependencies
 
 ```bash
-npm install
+pnpm install
 ```
 
 ## Storybook Setup
@@ -34,7 +34,7 @@ Runs storybook locally at `localhost:9010`
 ```shell
 cd dialtone-vue
 
-npm run storybook
+pnpm run storybook
 ```
 
 ### Building Storybook
@@ -44,7 +44,7 @@ Outputs the bundled storybook to `dialtone-vue/storybook-static`
 ```shell
 cd dialtone-vue
 
-npm run storybook:build
+pnpm run storybook:build
 ```
 
 ### Deploying Storybook
