+++ packages/dialtone-vue2/docs/storybook/writing_docs.mdx
@@ -65,7 +65,7 @@ documentation.
 <Controls />
 ```
 
-Note: `PRIMARY_STORY` is imported from `@storybook/addon-docs/blocks`.
+Note: `PRIMARY_STORY` is imported from `@storybook/addon-docs`.
 
 ### [Slots](https://v3.vuejs.org/guide/component-slots.html)
