+++ packages/dialtone-vue2/docs/storybook/writing_storybook.mdx
@@ -1,4 +1,4 @@
-import { Meta, Subtitle, Story, Canvas } from '@storybook/addon-docs/blocks';
+import { Meta, Subtitle, Story, Canvas } from '@storybook/addon-docs';
 import { DocImg } from '../components'
 import * as InputStories from '../../components/input/input.stories';
