+++ packages/dialtone-vue2/project.json
@@ -41,6 +41,13 @@
     },
     "release": {
       "executor": "nx:run-commands",
+      "dependsOn": [
+        {
+          "projects": [ "dialtone-icons" ],
+          "target": "build",
+          "params": "ignore"
+        }
+      ],
       "outputs": [],
       "options": {
         "command": "pnpm publish --filter ./packages/dialtone-vue2",
@@ -49,6 +56,13 @@
     },
     "release-github": {
       "executor": "nx:run-commands",
+      "dependsOn": [
+        {
+          "projects": [ "dialtone-icons" ],
+          "target": "build",
+          "params": "ignore"
+        }
+      ],
       "outputs": [],
       "options": {
         "command": "pnpm semantic-release-plus --extends ./packages/dialtone-vue2/release-ci.config.cjs",
