+++ packages/dialtone-vue2/recipes/comboboxes/combobox_multi_select/combobox_multi_select_story_constants.js
@@ -23,22 +23,3 @@ export const ITEMS_LIST_DATA = [
   { id: 'item21', value: 'item21', type: 'MAINLINE' },
   { id: 'item22', value: 'item22', type: 'Other' },
 ];
-
-export const MULTI_SELECT_SIZES = {
-  // Chip has no 'lg' and 'xl' size. So we don't support that in multi-select.
-  EXTRA_SMALL: 'xs',
-  SMALL: 'sm',
-  DEFAULT: 'md',
-};
-
-export const CHIP_SIZES = {
-  xs: 'xs',
-  sm: 'xs',
-  md: 'sm',
-};
-
-export const CHIP_TOP_POSITION = {
-  xs: 1.4,
-  sm: 0.4,
-  md: 0.2,
-};
