+++ packages/dialtone-vue2/recipes/comboboxes/combobox_multi_select/index.js
@@ -1,2 +1,2 @@
 export { default as DtRecipeComboboxMultiSelect } from './combobox_multi_select.vue';
-export { MULTI_SELECT_SIZES } from './combobox_multi_select_story_constants';
+export { MULTI_SELECT_SIZES, CHIP_SIZES, CHIP_TOP_POSITION } from './combobox_multi_select_constants.js';
