+++ packages/dialtone-vue3/docs/templates/story_template.mdx
@@ -113,7 +113,6 @@ const DefaultTemplate = (args, { argTypes }) => createTemplateFromVueFile(
 const VariantsTemplate = (args, { argTypes }) => createTemplateFromVueFile(
   args,
   argTypes,
-  argTypes,
   MyComponentVariantsTemplate,
 );
