+++ packages/dialtone-vue3/recipes/comboboxes/combobox_multi_select/combobox_multi_select.vue
@@ -125,8 +125,8 @@ import {
   MULTI_SELECT_SIZES,
   CHIP_SIZES,
   CHIP_TOP_POSITION,
-} from './combobox_multi_select_story_constants';
-import SrOnlyCloseButtonMixin from '../../../common/mixins/sr_only_close_button';
+} from './combobox_multi_select_constants';
+import SrOnlyCloseButtonMixin from '@/common/mixins/sr_only_close_button';
 
 export default {
   name: 'DtRecipeComboboxMultiSelect',
