--- packages/dialtone-vue3/scripts/commit-msg
@@ -1,8 +1,0 @@
-#!/bin/sh
-
-branch="$(git rev-parse --abbrev-ref HEAD)"
-
-# Validate commit message in production and staging branches
-if [[ "$branch" = "staging" || "$branch" = "production" ]]; then
-  npx --no-install commitlint --edit "$1"
-fi
